msggen: Move rpc-model generator
