+++ contrib/msggen/msggen/gen/__init__.py
@@ -1,3 +1,3 @@
 from .generator import IGenerator, GeneratorChain # noqa
 from .grpc import GrpcGenerator, GrpcConverterGenerator, GrpcUnconverterGenerator, GrpcServerGenerator # noqa
-from .rust import RustGenerator # noqa
+from .rpc.rust import RustGenerator # noqa
