+++ contrib/msggen/msggen/gen/rpc/rust.py
